pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "continuum-rl-obstacle-avoidance"
version = "0.1.0"
description = "RL-based obstacle avoidance control for a soft continuum robot"
readme = "README.md"
requires-python = ">=3.9,<3.10"
dependencies = [
  "numpy==1.22.3",
  "matplotlib==3.5.2",
  "pandas==1.4.4",
  "tensorflow==2.13.0",
  "torch==2.0.0",
  "protobuf==3.20.3",
  "gym==0.26.2",
  "ipython==8.11.0",
  "hydra-core==1.3.2",
  "omegaconf==2.3.0",
]

[project.optional-dependencies]
wandb = [
  "wandb==0.16.6",
]
dev = [
  "pytest==9.0.2",
  "ruff==0.15.5",
]

[project.scripts]
continuum-rl = "continuum_rl.hydra_app:main"
continuum-rl-compat = "continuum_rl.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = [
  "continuum_rl*",
  "continuum_robot*",
  "kinematics*",
  "Pytorch*",
  "Keras*",
]

[tool.setuptools.package-data]
"continuum_rl.conf" = ["*.yaml", "task/*.yaml"]
"continuum_rl.data" = ["*.txt", "*.npy"]

[tool.continuum_rl.exports]
conda_env_name = "continuum-rl"
conda_python = "3.9.7"
